refactor: - rename and update auth session files - auth api interface by correct response
