--- data/src/main/java/com/london/data/datasource/remote/auth/model/CreateSessionRequest.kt
@@ -1,5 +1,0 @@
-package com.london.data.datasource.remote.auth.model
-
-data class CreateSessionRequest(
-    val requestToken: String
-)
