+++ data/src/main/java/com/london/data/datasource/remote/auth/model/CreateSessionWithLoginRequest.kt
@@ -1,0 +1,12 @@
+package com.london.data.datasource.remote.auth.model
+
+import kotlinx.serialization.SerialName
+
+data class CreateSessionWithLoginRequest(
+    @SerialName("username")
+    val username: String,
+    @SerialName("password")
+    val password: String,
+    @SerialName("request_token")
+    val requestToken: String
+)
